+++ Gomoku.CLI/Gomoku.CLI.csproj
@@ -5,6 +5,7 @@
         <TargetFramework>net8.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
+        <RootNamespace>Gomoku.CLI</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
