+++ Gomoku.CLI/PerformanceTest.cs
@@ -1,0 +1,6 @@
+﻿namespace Gomoku.CLI;
+
+public class PerformanceTest
+{
+    
+}
