+++ Gomoku.CLI/Program.cs
@@ -1,0 +1,4 @@
+﻿using Gomoku.CLI;
+
+QualityComparison.CompareQuality();
+
